requirements.txt
mcp[cli]>=1.2.0
requests==2.31.0
python-dotenv==1.0.1